Flatten docs home component
